backend/requirements.txt
langgraph
langchain-core
langchain-groq
langchain
langchain-community
requests
fastapi
pydantic
python-dotenv
faiss-cpu
sentence-transformers
pypdf
uvicorn